requirements.txt
(Flask==2.2.5)
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Mail==0.9.1
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
WTForms==3.0.1
click>=8.0
alembic>=1.11
python-dotenv
